docs and various tweaks
