+++ target_hubspot/context.py
@@ -1,10 +1,11 @@
 from logging import Logger
 
-from target_hubspot.constants import TargetConfig
+from target_hubspot.model import TargetConfig
 
 
-class ConfigInheriter:
+class JobContext:
     # We tend to pass a few objects around to most classes, so may as well standardize it properly w/ inheritance
+    # Classes looking to use this simply have to inherit it and make a call to super().__init__(config=config, logger=logger)
     _config: TargetConfig
     _logger: Logger
