--- target_hubspot/exception.py
@@ -1,7 +1,0 @@
-"""Custom exceptions."""
-
-
-class PartialImportException(Exception):
-    """Raised when an import has errors."""
-
-    pass
