+++ target_hubspot/model.py
@@ -1,5 +1,6 @@
 from enum import Enum
-from typing import Any, List, Literal, Optional
+
+from pydantic import Field
 
 import target_hubspot.pydantic_config
 
@@ -9,19 +10,6 @@ class HubspotObjectsEnum(str, Enum):
     CONTACTS = "contacts"
     COMPANIES = "companies"
     DEALS = "deals"
-
-
-class GetNewToken:
-    class RequestPayload(target_hubspot.pydantic_config.BaseConfig):
-        grant_type: Literal["refresh_token"] = "refresh_token"
-        client_id: str
-        client_secret: str
-        refresh_token: str
-
-    class ResponseBody(target_hubspot.pydantic_config.BaseConfig):
-        access_token: str
-        refresh_token: str
-        expires_in: int
 
 
 class HubspotDataTypes(str, Enum):
@@ -46,42 +34,29 @@ class HubspotFieldTypes(str, Enum):
     CALCULATION_EQUATIUON = "calculation_equation"
 
 
-class HubspotPropertyPayload(target_hubspot.pydantic_config.BaseConfig):
-    # Derived from: https://developers.hubspot.com/docs/api/crm/properties
-    id: str
-    hidden: Optional[bool] = None
-    displayOrder: Optional[int] = None
-    description: Optional[str] = None
-    label: str
-    type: HubspotDataTypes
-    formField: Optional[bool] = None
-    groupName: str
-    referencedObjectType: Optional[str] = None
-    name: str
-    options: Optional[List[str]] = None
-    calculationFormula: Optional[str] = None
-    hasUniqueValue: Optional[bool] = None
-    fieldType: HubspotFieldTypes
-    externalOptions: Optional[bool] = None
+class TargetConfig(target_hubspot.pydantic_config.BaseConfig):
+    # We define our config both here and in the main `target.py` file because the SDK checks at runtime but does not get you mypy-level type safety like we want
+    client_id: str = Field(
+        ...,
+        min_length=1,
+        description="Our OAuth app client ID.",
+    )
+    client_secret: str = Field(
+        ...,
+        min_length=1,
+        description="Our OAuth app client secret. Persistent across all installs, but we still allow it to be injected.",
+    )
+    refresh_token: str = Field(
+        ...,
+        min_length=1,
+        description="The refresh token of the current user's OAuth connection.",
+    )
 
-
-class BatchCreateProperties:
-    class RequestPayload(target_hubspot.pydantic_config.BaseConfig):
-        inputs: List[HubspotPropertyPayload]
-
-class HubspotContactUpdatePayload(target_hubspot.pydantic_config.BaseConfig):
-    id: int # MUST be the basic hubspot id; they claim in the docs you can use email and supply idProperty=email, but the bulk update endpoint just straight-up doesn't support it and it's horribly undocumented
-    properties: dict[str, Any]
-
-class BatchUpdateContacts:
-    class RequestPayload(target_hubspot.pydantic_config.BaseConfig):
-        inputs: List[HubspotContactUpdatePayload]
-
-    class ResponseBody(target_hubspot.pydantic_config.BaseConfig):
-        pass
-
-class CreatePropertyGroup:
-
-    class RequestPayload(target_hubspot.pydantic_config.BaseConfig):
-        name: str
-        label: str
+    object_type: HubspotObjectsEnum = Field(
+        ...,
+        description=f"The object type to upload data to. Supported types: {HubspotObjectsEnum.__members__.values()}",
+    )
+    filepath: str = Field(
+        ...,
+        description="The file path to upload data from.",
+    )
