+++ tests/resources/companies.jsonl
@@ -1,0 +1,4 @@
+{"type": "SCHEMA", "stream": "target-hubspot_test-test_simple_table", "schema": {"definitions": {"sdc_recursive_boolean_array": {"items": {"$ref": "#/definitions/sdc_recursive_boolean_array"}, "type": ["null", "boolean", "array"]}, "sdc_recursive_integer_array": {"items": {"$ref": "#/definitions/sdc_recursive_integer_array"}, "type": ["null", "integer", "array"]}, "sdc_recursive_number_array": {"items": {"$ref": "#/definitions/sdc_recursive_number_array"}, "type": ["null", "number", "array"]}, "sdc_recursive_object_array": {"items": {"$ref": "#/definitions/sdc_recursive_object_array"}, "type": ["null", "object", "array"]}, "sdc_recursive_string_array": {"items": {"$ref": "#/definitions/sdc_recursive_string_array"}, "type": ["null", "string", "array"]}, "sdc_recursive_timestamp_array": {"format": "date-time", "items": {"$ref": "#/definitions/sdc_recursive_timestamp_array"}, "type": ["null", "string", "array"]}}, "properties": {"id": {"type": ["integer"]}, "akkio_predicted_value": {"maximum": 1, "minimum": 0, "type": ["number"]}}, "type": "object"}, "key_properties": ["id"], "bookmark_properties": ["id"]}
+{"type": "RECORD", "stream": "target-hubspot_test-test_simple_table", "record": {"id": 5922, "akkio_predicted_value": 0.8}, "version": 1, "time_extracted": "2023-11-01T19:12:12.006049Z"}
+{"type": "RECORD", "stream": "target-hubspot_test-test_simple_table", "record": {"id":  461, "akkio_predicted_value": 0.3}, "version": 1, "time_extracted": "2023-11-01T19:12:12.006049Z"}
+{ "type": "STATE", "value": { "id": 2 }}
